% Auto-generated: do not edit by hand
\name{gQLClient}

\alias{gQLClient}

\title{GQLClient component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
gQLClient(id=NULL, data=NULL, error=NULL, loading=NULL, query=NULL,
uri=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{Named list. The value displayed in the input.}

\item{error}{Named list | character. }

\item{loading}{Logical. }

\item{query}{Character. A label that will be printed when this component is rendered.}

\item{uri}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

